+++ UI/_base/HTML.ts
@@ -18,34 +18,15 @@ import AppData from './AppData';
 import { IHTMLOptions } from './interface/IHTML';
 import { IRootTemplateOptions } from './interface/IRootTemplate';
 import { headDataStore } from 'UI/_base/HeadData';
+import mountChecker from 'UI/_base/MountChecker';
 
 // Бывают ситуации, когда страницу открыли и сразу перешли на другую вкладку или перевели компьютер в режим сна.
 // У открытой страницы в фоновом режиме начинают по таймауту отваливаться запросы и страница в итоге не оживает.
-// Для обработки таких ситуаций запустим проверку в интервале. Запускаем именно интервал, а не таймаут на максимально
-// допустимое время, т.к. если компьютер перейдет в режим сна, то оставшееся время таймаута сохранится.
-// И в таком случае после пробуждения страница перезагрузится не сразу, а когда пройдет оставшееся время таймаута.
-// Если шаблон страницы построился, то в _afterMount мы остановим проверку, т.к. страница рабочая и проверять
-// что-то больше не надо. Если же мы попали в обработчик, занчит страница еще не ожила.
-// Проврим что время от начала загрузки прошло достаточно, чтобы страница построилась. Если же страница не
-// построилась за максимально допустимое время, нужно ее перезагрузить. Интервал запускается сразу при загрузке файла,
-// т.к. если построение страницы упадет, _beforeMount и _afterMount могут не выполниться.
-let reloadPageInterval;
-if (constants.isBrowserPlatform) {
-    // Интервал запускаем только на новых страницах, т.к. этот файл может прилететь в пакете на старые страницы и запустить
-    // интервал. Но т.к. файл грузится просто как зависимость, то интервал никто не удалит и страница обновится.
-    const isNewEnvironment = !!document.getElementsByClassName('ui-HTML').length;
-
-    if (isNewEnvironment) {
-        const startPageLoadTime = Date.now();
-        const PAGE_CHECK_INTERVAL = 2 * 1000;
-        const MAX_PAGE_MOUNT_TIME = 2 * 60 * 1000;
-        reloadPageInterval = setInterval(() => {
-            if (Date.now() - startPageLoadTime > MAX_PAGE_MOUNT_TIME) {
-                window.location.reload();
-            }
-        }, PAGE_CHECK_INTERVAL);
-    }
-}
+// Для обработки таких ситуаций запустим скрипт, который перезагрузит страницу, если она не оживет за отведенное время.
+// Проверку запускается сразу при загрузке файла, т.к. если построение страницы упадет, _beforeMount может не
+// выполниться. Если шаблон страницы построился, то в _afterMount мы остановим проверку, т.к. страница рабочая и
+// проверять что-то больше не надо.
+mountChecker.start();
 
 interface IHTMLCombinedOptions extends IHTMLOptions, IRootTemplateOptions {
     // Добавим здесь поля для RUM-статистики Потому что их нам нужно сериализовать в wsConfig, чтобы потом получить на клиенте.
@@ -173,7 +154,7 @@ class HTML extends Control {
     }
 
     _afterMount(): void {
-        clearInterval(reloadPageInterval);
+        mountChecker.stop();
         function inIframe() {
             try {
                 return window.self !== window.top;
